上线之前的修改

+++ 30000day/3.Shop/comment/view/ShopDetailCommentTableViewCell.m
@@ -130,7 +130,7 @@
             
         }
         
-        self.commentCountLable.text = [NSString stringWithFormat:@"%ld",informationCommentModel.clickLikeCount];
+        self.commentCountLable.text = [NSString stringWithFormat:@"%d",(int)informationCommentModel.clickLikeCount];
     }
